Support for managing secrets with AKV key vaults (#12)
* Initial support for importing and releasing RSA keys from key vault

* Oct key derivation from RSA key

* Encfs support for deriving oct key from RSA keys

* Importkey reads PEM file. SKR sidecar response carries a key in JWK format. KeyDerivation metadata passed as a seperate blob

* Updated README files

* updated README

* Initial support for importing and releasing RSA keys from key vault

* Oct key derivation from RSA key

* Encfs support for deriving oct key from RSA keys

* Importkey reads PEM file. SKR sidecar response carries a key in JWK format. KeyDerivation metadata passed as a seperate blob

* changes to tech-info

* updated README

* Fixed api bugs after rebasing with main

* Minor fixes

* mhsm -> akv change

* Removed println

* Missing files

* Removed hmsm references from techinfo

* Bump github.com/gin-gonic/gin from 1.7.4 to 1.7.7 (#11)

Bumps [github.com/gin-gonic/gin](https://github.com/gin-gonic/gin) from 1.7.4 to 1.7.7.
- [Release notes](https://github.com/gin-gonic/gin/releases)
- [Changelog](https://github.com/gin-gonic/gin/blob/master/CHANGELOG.md)
- [Commits](https://github.com/gin-gonic/gin/compare/v1.7.4...v1.7.7)

---
updated-dependencies:
- dependency-name: github.com/gin-gonic/gin
  dependency-type: direct:production
...


* Initial support for importing and releasing RSA keys from key vault

* Oct key derivation from RSA key

* Encfs support for deriving oct key from RSA keys

* Importkey reads PEM file. SKR sidecar response carries a key in JWK format. KeyDerivation metadata passed as a seperate blob

* changes to tech-info

* updated README

* Initial support for importing and releasing RSA keys from key vault

* Oct key derivation from RSA key

* Encfs support for deriving oct key from RSA keys

* Updated README files

* Fixed api bugs after rebasing with main

* Minor fixes

* mhsm -> akv change

* Removed println

* Missing files

* Removed hmsm references from techinfo

* Consistent importkey configs. Removed sidecar-args file and from arm template

* Changed the SecureKeyRelease API to return JWK key rather than raw bytes and key type

* Comments

* Arm template tags

---------

+++ examples/encfs/generatefs/generatefs.sh
@@ -9,15 +9,18 @@ encryptedImage=encfs.img
 deviceName=cryptdevice1
 deviceNamePath="/dev/mapper/$deviceName"
 
-echo "[!] Generating keyfile..."
-
-dd if=/dev/random of="$keyFilePath" count=1 bs=32
+if [ -f "$keyFilePath" ]; then
+    echo "keyfile exists"
+else
+    echo "[!] Generating keyfile..."
+    dd if=/dev/random of="$keyFilePath" count=1 bs=32
+fi
 
 echo "Key in hex string format"
 
 python hexstring.py $keyFilePath
 
-truncate -s 64 "$keyFilePath"
+truncate -s 32 "$keyFilePath"
 
 echo "[!] Creating encrypted image..."
